==> scripts/6_PrevalenceAnalyses.R <==
# Analysis scripts for Baez-Ortega et al., 2019
# Step 6. Somatic mutation prevalence estimation

# Adrian Baez-Ortega, 2018


# TO RUN THIS SCRIPT IN THE TERMINAL
# ----------------------------------
# Run the commands below in the terminal, replacing '/path/to/TCG2019' with 
# the path to the TCG2019 directory.
#
#    cd /path/to/TCG2019
#    Rscript scripts/6_PrevalenceAnalyses.R


# TO RUN THIS SCRIPT IN RSTUDIO
# -----------------------------
# Before starting, run the line below in RStudio, replacing '/path/to/TCG2019' with 
# the path to the TCG2019 directory.
#
#    setwd("path/to/TCG2019")


# If the paths to the input or output files differ from the ones 
# used in this script, they may be updated by modifying the lines below.

INPUT = list(
    
    # Path to input file containing human cancer mutation counts from Alexandrov et al., 2013 (doi.org/10.1038/nature12477)
    # (from exome mutation data available at ftp://ftp.sanger.ac.uk/pub/cancer/AlexandrovEtAl/mutational_catalogs/exomes)
    MUT.COUNTS = file.path("data", "original", "AlexandrovEtAl_Human_Mutation_Counts.tsv"),
    
    # Path to input file generated by script 2_ImportVariants.R
    VAR.TABLES = file.path("data", "processed", "Variant_Tables.RData"),
    
    # Path to input file generated by script 5_PhyloGroups.R
    PHYLO.GROUPS = file.path("data", "processed", "Phylo_Tree_Groups.RData")
    
)

OUTPUT = list(
    
    # Path to output PDF file
    PLOT.PDF = file.path("output", "Somatic_Mutation_Prevalence.pdf")
    
)


# Constants
DOG.EXOME.LEN = 43.64 * 1e6  # 43,639,327 bp sequenced
HUMAN.EXOME.LEN = 30 * 1e6   # Exome size used in Alexandrov et al., 2013 (doi.org/10.1038/nature12477)


# Print input and output file paths
cat("\nInput files:")
for (name in INPUT) {
    cat("\n  ", name)
}
cat("\n\nOutput files:")
for (name in OUTPUT) {
    cat("\n  ", name)
}
cat("\n\n")


# Load input data
cat("Loading data...\n")
human.counts = read.table(INPUT$MUT.COUNTS, sep="\t", header=T, stringsAsFactors=F)
load(INPUT$VAR.TABLES)
load(INPUT$PHYLO.GROUPS)


# Calculate mutation prevalence in human cancers
tum.types = unique(human.counts$Type)
human.preval = sapply(tum.types, function(type) {
    human.counts$Mutations[human.counts$Type == type] / HUMAN.EXOME.LEN * 1e6
})
human.preval.median = sapply(human.preval, median)
human.preval = human.preval[order(human.preval.median)]
human.preval = lapply(human.preval, sort)


# Calculate mutation prevalence in CTVT
ctvt.counts = sort(colSums(snvs.nv[tumour.only.idx, tumours.tree] >= MIN.READS))
ctvt.preval = ctvt.counts / DOG.EXOME.LEN * 1e6


# Plot prevalence for all human cancer types
cat("\nPlotting mutation prevalence plots to output directory...\n")

pdf(OUTPUT$PLOT.PDF, 8, 4, onefile=T)
par(mar=c(5.5, 5, 1.5, 1))

OFFSET = 50
limits = c(0, lengths(human.preval))
for (i in 2:length(limits)) {
    limits[i] = limits[i] + limits[i-1]
}
limits[1] = limits[1] - OFFSET
limits[length(limits)] = limits[length(limits)] + OFFSET
centres = sapply(2:length(limits), function(i) {
    mean(limits[(i-1):i])
})

TOP = 2.815
TOP2 = 2.99
BOTTOM = -1.815
BOTTOM2 = -1.99
LEFT = limits[1]
RIGHT = limits[length(limits)]
MEDLEN = 80

plot(log10(unlist(human.preval)), 
     ylim=c(BOTTOM, TOP), xlim=c(LEFT, RIGHT),
     yaxt="n", xaxt="n", xlab="", ylab="", xaxs="i", col=NA)

for (i in seq(1, length(limits), 2)) {
    polygon(x=limits[c(i, i+1, i+1, i)], y=c(rep(BOTTOM2, 2), rep(TOP2, 2)),
            col="grey90", border=NA)
}
abline(h=-1:2, lty=3, lwd=0.5)
axis(side=2, at=-2:3, labels=10^(-2:3), las=2, cex.axis=0.8)
mtext(side=2, text="Somatic SNV prevalence\n(SNVs per Mb)", line=3, cex=0.8)

segments(x0=centres - MEDLEN, x1=centres + MEDLEN, 
         y0=log10(sort(human.preval.median)), 
         col="firebrick3", lwd=2.5)
points(log10(unlist(human.preval)), pch=16, 
       ylim=c(BOTTOM, TOP), xlim=c(-10, limits[length(limits)] + 10),
       xaxs="i", cex=0.4)

mtext(side=1, at=centres, text=tum.types[order(human.preval.median)], 
      adj=1, las=2, line=0.2, cex=0.6)


# Plot prevalence for CTVT and selected human types:
# ALL, Neuroblastoma, Prostate, Head and Neck, Lung Squamous, Melanoma
par(mfrow=c(1, 7), mar=c(4, 0, 5.5, 0), oma=c(0, 6.5, 0, 1))

TOP = 3.2
BOTTOM = -2.2
T.CEX = 0.87
T.LINE = -0.7
P.CEX = 0.9
TYPES = c("ALL", "Neuroblastoma", "Prostate", "Head and Neck", "Lung Squamous", "Melanoma")

# Human cancers
for (i in 1:length(TYPES)) {
    type = TYPES[i]
    
    RIGHT = length(human.preval[[type]]) * 1.05
    LEFT = length(human.preval[[type]]) - RIGHT
    CENTRE = (LEFT + RIGHT) / 2
    
    plot(log10((human.preval[[type]])), 
         ylim=c(BOTTOM, TOP), xlim=c(LEFT, RIGHT), 
         pch=16, xlab="", ylab="", type="n", axes=F)
    polygon(x=c(LEFT, RIGHT, RIGHT, LEFT) * 1.1,
            y=c(BOTTOM, BOTTOM, TOP, TOP),
            col=ifelse(i %% 2 == 0, "grey93", NA), border=NA)
    abline(h=-2:3, lty=3, col="gray60")
    
    segments(x0=CENTRE * 0.3, x1=CENTRE * 1.77, 
             y0=log10(median(human.preval[[type]])), 
             col="red3", lwd=4)
    points(log10((human.preval[[type]])),
           ylim=c(BOTTOM, TOP), xlim=c(LEFT, RIGHT),
           pch=16, cex=P.CEX, xlab="", ylab="")
    
    mtext(side=3, line=T.LINE, cex=T.CEX, at=CENTRE * 1.04,
          text=ifelse(i == 1, type,
                      paste0(substr(type, 1, 1), 
                             tolower(substr(type, 2, nchar(type))))))
    
    abline(h=c(BOTTOM, TOP), xpd=T)
    if (i == 1) {
        axis(side=2, at=-2:3, labels=10^(-2:3), las=2, cex.axis=1.2, line=-0.02)
        mtext(side=2, text="Somatic SNV prevalence\n(SNVs per Mb)", line=3.5, cex=0.9)
        segments(x0=-12.5, y0=BOTTOM, y1=TOP, xpd=T)
    }
}

# CTVT
RIGHT = length(ctvt.preval) * 1.03
CENTRE = (LEFT + RIGHT) / 2

plot(log10(ctvt.preval),
     ylim=c(BOTTOM, TOP), xlim=c(LEFT, RIGHT),
     type="n", pch=16, cex=P.CEX, axes=F, xlab="", ylab="")

segments(x0=CENTRE * 0.3, x1=CENTRE * 1.77,
         y0=log10(median(ctvt.preval)),
         col="red3", lwd=5)
points(log10(ctvt.preval),
       ylim=c(BOTTOM, TOP), xlim=c(LEFT, RIGHT),
       pch=16, cex=P.CEX, xlab="", ylab="")

abline(h=-2:3, lty=3, col="gray60")
mtext(side=3, line=T.LINE, cex=T.CEX, text="CTVT", at=CENTRE * 1.04)
abline(h=c(BOTTOM, TOP), xpd=T)
segments(x0=RIGHT * 1.037, y0=BOTTOM, y1=TOP, xpd=T)

invisible(dev.off())

cat("\nDone\n\n")
